Update patch set 1
Patch Set 1:

The commit is replacing python-retrying with python-tenacity.

This is not affecting the ocata-uc worker because it is still based on a tagged release before this was introduced (5.12.0 doesn't include it, so it will be 5.13.0).

I'll propose a fix, but I guess it should wait to be landed until a 5.13.0 release is tagged.

Patch-set: 1
Label: Code-Review=0